requirements.txt
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
lxml==6.0.2
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0

# Novas dependências
pydantic==2.7.1          # Validação de dados (compatível com Prefect 3.0)
prefect==3.0.0           # Agendamento e orquestração
sqlalchemy==2.0.23       # ORM (opcional, para futuro)
python-dotenv==1.1.0     # Variáveis de ambiente
psycopg2-binary==2.9.9   # Driver PostgreSQL
streamlit==1.32.2        # Dashboard web
pandas==2.1.4            # Análise de dados
plotly==5.18.0           # Gráficos interativos